requirements.txt
numpy==1.17.0
Pillow==6.1.0
pyspark==2.3.0
wheel==0.26.0
